requirements.txt
-f /usr/share/pip-wheels
arabic-reshaper==3.0.0
asgiref==3.5.2
asn1crypto==1.5.1
backports.zoneinfo==0.2.1
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==2.1.1
click==8.1.7
cryptography==43.0.0
cssselect2==0.7.0
Django==4.1
django-filter==22.1
html5lib==1.1
idna==3.7
lxml==5.2.2
oscrypto==1.3.0
Pillow==9.2.0
pycparser==2.22
pyHanko==0.25.1
pyhanko-certvalidator==0.26.3
PyPDF3==1.0.6
pypng==0.20220715.0
python-bidi==0.4.2
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0.1
qrcode==7.4.2
reportlab==3.6.11
requests==2.31.0
six==1.16.0
sqlparse==0.4.2
svglib==1.5.1
tinycss2==1.3.0
tqdm==4.66.4
typing-extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
uritools==4.0.3
urllib3==1.26.11
webencodings==0.5.1
xhtml2pdf==0.2.8
